feat : Cell 세부 내용 수정

+++ RankIn/RankIn/Domain/Model/RankTableViewCellContents.swift
@@ -11,6 +11,6 @@ struct RankTableViewCellContents: Hashable {
     
     let rank: Int
     let nickName: String
-    let score: Int
+    let score: Double
     
 }
